Level Design + Prophet + Animation
This includes the updated level design as it is now. There is also the
game ready version of Prophet that is playable  and some of her
temporary animations. Such as Run, Idle, One Melee Attack, and a couple
others. The textures used for the landscape are Substances, so please
download the Substance Plug In from the Unreal Marketplace otherwise the
textures may not work right...Some properties to change within the plug
in is just the amount of resolution allowed, it's set i think by default
to 256x256, 1024 or 2048 are better options.
I hope that is everything...and i hope i did this right.